requirements.txt
aniso8601==9.0.1
attrs==21.4.0
click==8.1.2
Flask==2.1.1
Flask-Cors==3.0.10
flask-restx==0.5.1
importlib-metadata==4.11.3
importlib-resources==5.6.0
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
Werkzeug==2.1.1
zipp==3.8.0
